Backend/requirements.txt
colorama==0.4.6
fastapi==0.124.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
uvicorn==0.38.0